Added error message
Sadly I am dumb, it was transitioning from a stack loop to an interior loop which assumed the original pairing was a binding. If you incorrectly specify a structure that has invalid bindings it can cause an illegal array access that could have any number of problems. This error checks the value before indexing