Patch fix for -1 indexing on openloop. Corrected optimization bug
Patch fix for -1 indexing on openloop. I created a wrapper array for thisCodeSeq. The problem here was that when optimized there was no guarantee that the block of memory before the array wasn't going to be moved around. With the wrapper the -1 block will be treated as the same block of mem. This is basically a patch but we all know that those are permanent. Also the generateLoops function was not returning a value when not compiled in -O0. Not sure why zero optimizations didn't catch this, nor the compiler but oh well.